backend/requirements.txt
flask
flask-cors
cryptography
google-api-python-client
google-auth-oauthlib
supabase